refact (Swift): created commons module with every reusable object
