+++ commons/src/main/java/fr/insideapp/sonaqube/apple/commons/issues/ReportIssue.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonarqube.swift.lang.issues;
+package fr.insideapp.sonaqube.apple.commons.issues;
 
 import java.util.Objects;
