+++ scripts/commons/ConsoleString.groovy
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package common
+package commons
 /**
  * A String category to enable console output coloring / formatting.
  */
