+++ scripts/commons/RuleUpdater.groovy
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package common
+package commons
 
 import groovy.json.JsonSlurper
 import groovy.json.JsonBuilder
