+++ scripts/updateSwiftLintRules.groovy
@@ -15,8 +15,8 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import common.RuleUpdater
-import common.ConsoleString
+import commons.RuleUpdater
+import commons.ConsoleString
 
 String.mixin(ConsoleString)
