+++ swift-lang/src/main/java/fr/insideapp/sonarqube/swift/lang/antlr/AntlrContext.java
@@ -17,8 +17,8 @@
  */
 package fr.insideapp.sonarqube.swift.lang.antlr;
 
-import fr.insideapp.sonarqube.swift.lang.SourceLine;
-import fr.insideapp.sonarqube.swift.lang.SourceLinesProvider;
+import fr.insideapp.sonaqube.apple.commons.SourceLine;
+import fr.insideapp.sonaqube.apple.commons.SourceLinesProvider;
 import fr.insideapp.sonarqube.swift.lang.antlr.generated.Swift5Lexer;
 import fr.insideapp.sonarqube.swift.lang.antlr.generated.Swift5Parser;
 import org.antlr.v4.runtime.CharStream;
