+++ swift-lang/src/main/java/fr/insideapp/sonarqube/swift/lang/issues/swiftlint/SwiftLintRulesDefinition.java
@@ -28,7 +28,6 @@ import org.sonar.api.utils.log.Logger;
 import org.sonar.api.utils.log.Loggers;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 public class SwiftLintRulesDefinition implements RulesDefinition {
